All parsers now set the year, if the parsed year is 1970 or earlier
